fixed route problem with getAll

+++ src/modules/session/model.js
@@ -3,7 +3,7 @@ const db = require("../../config/db-config");
 const findAll = () => {
   return db
     .execute(
-      "SELECT *, session.id FROM session JOIN location ON session.location_id=location.id ORDER BY session.date ASC"
+      "SELECT *, session.id, lat, lng  FROM session JOIN location ON session.location_id=location.id ORDER BY session.date ASC"
     )
     .then((data) => {
       return data;
@@ -20,11 +20,9 @@ const findAllWithNumberOfParticipants = () => {
 };
 
 const findById = (id) => {
-  return db
-    .execute("SELECT * FROM session WHERE id = ? ", [id])
-    .then(([data]) => {
-      return data;
-    });
+  return db.execute("SELECT * FROM session WHERE id = ? ", [id]).then(([data]) => {
+    return data;
+  });
 };
 
 const findByIdWithNumberOfParticipants = (id) => {
@@ -40,10 +38,9 @@ const findByIdWithNumberOfParticipants = (id) => {
 
 const findWineBySessionId = (id) => {
   return db
-    .execute(
-      "SELECT * FROM session_has_wine JOIN wine ON wine.id = session_has_wine.wine_id WHERE session_id = ?",
-      [id]
-    )
+    .execute("SELECT * FROM session_has_wine JOIN wine ON wine.id = session_has_wine.wine_id WHERE session_id = ?", [
+      id,
+    ])
     .then(([data]) => {
       return data;
     });
@@ -61,10 +58,7 @@ const findUserBySessionId = (id) => {
 };
 
 const registerSession = (user_id, session_id) => {
-  return db.execute(
-    `INSERT INTO session_has_user (user_id, session_id) VALUES (?,?)`,
-    [user_id, session_id]
-  );
+  return db.execute(`INSERT INTO session_has_user (user_id, session_id) VALUES (?,?)`, [user_id, session_id]);
 };
 
 const createNewSession = (session) => {
@@ -82,10 +76,7 @@ const createUserHasSession = (data) => {
 
 const createWineHasSession = (data) => {
   const { sessionId, wineId } = data;
-  return db.execute(
-    "INSERT INTO session_has_wine (wine_id, session_id) VALUES (?, ?)",
-    [wineId, sessionId]
-  );
+  return db.execute("INSERT INTO session_has_wine (wine_id, session_id) VALUES (?, ?)", [wineId, sessionId]);
 };
 
 const deleteSessionById = (id) => {
@@ -93,10 +84,7 @@ const deleteSessionById = (id) => {
 };
 
 const getUserSessionsBySessionId = (user_id, session_id) => {
-  return db.execute(
-    "Select * from session_has_user where user_id = ? and session_id =?",
-    [user_id, session_id]
-  );
+  return db.execute("Select * from session_has_user where user_id = ? and session_id =?", [user_id, session_id]);
 };
 
 const findSessionByUserId = (id) => {
@@ -112,18 +100,12 @@ const findSessionByUserId = (id) => {
 
 const deleteUserFromSessionById = (data) => {
   const { sessionid, userid } = data;
-  return db.execute(
-    "DELETE FROM session_has_user WHERE session_id = ? AND user_id = ?",
-    [sessionid, userid]
-  );
+  return db.execute("DELETE FROM session_has_user WHERE session_id = ? AND user_id = ?", [sessionid, userid]);
 };
 
 const deleteWineFromSessionById = (data) => {
   const { sessionid, wineid } = data;
-  return db.execute(
-    "DELETE FROM session_has_wine WHERE session_id = ? AND wine_id = ?",
-    [sessionid, wineid]
-  );
+  return db.execute("DELETE FROM session_has_wine WHERE session_id = ? AND wine_id = ?", [sessionid, wineid]);
 };
 
 module.exports = {
